+++ src/ShippingFile/Views/Cnab240.php
@@ -61,9 +61,9 @@ abstract class Cnab240 extends BankInterchange\ShippingFile\View
     /**
      * Adds a Title registry
      *
-     * @param Models\ShippingFileTitle $sft Contains data for the registry
+     * @param Models\Title $title Contains data for the registry
      */
-    protected function add(Models\ShippingFileTitle $sft)
+    protected function add(Models\Title $title)
     {
         if ($this->current_lot === 99999) {
             $this->registry_count++;
@@ -84,7 +84,7 @@ abstract class Cnab240 extends BankInterchange\ShippingFile\View
 
         $this->registries = array_merge(
             $this->registries,
-            $this->generateLotDetail($sft)
+            $this->generateLotDetail($title)
         );
     }
 
@@ -126,13 +126,11 @@ abstract class Cnab240 extends BankInterchange\ShippingFile\View
     /**
      * Generates LotDetail registry
      *
-     * @param Models\ShippingFileTitle $sft Contains data for the registry
+     * @param Models\Title $title Contains data for the registry
      *
      * @return string[]
      */
-    abstract protected function generateLotDetail(
-        Models\ShippingFileTitle $sft
-    );
+    abstract protected function generateLotDetail(Models\Title $title);
 
     /**
      * Generates LotTrailer registry